backend/requirements.txt
fastapi
uvicorn[standard]
redisvl
python-multipart
python-dotenv
pydantic
pydantic-settings
openai
sentence-transformers
PyPDF2
python-docx
aiofiles
httpx
